MalecPiotr_StolarekSzymon/weather_app/requirements.txt
matplotlib
requests
mysql-connector-python
coloredlogs
logging
tkinter